// packages/backend/src/modules/1_mos_connection/connector/api.ts
export interface IConnectionConfig {
	/** The ID of this mos-device */
	mosID: string
	/** Whether this mosConnection accepts new connections from othe MOS clients (ie acts as an NCS) */
	acceptsConnections: boolean
	/** Only accept connections from this whitelist */
	accepsConnectionsFrom?: string[]
	/** A list of which profile this mos device is to support */
	profiles: IProfiles
	/** If true, this device is assumed to be an NCS (server). Defaults to a MOS (client). */
	isNCS?: boolean
	/** Debugging-mode: logs raw mos-messages */
	debug?: boolean
	offspecFailover?: boolean
	/** If set to true, a strict check is performed to ensure that all required callbacks are set up for specified profiles */
	strict?: boolean
	/** If set, overrides the standard port numbers */
	ports?: {
		/** Set MOS Lower port (standard: 10540) */
		lower: number
		/** Set MOS Upper port (standard: 10541) */
		upper: number
		/** Set MOS Query port (standard: 10542) */
		query: number
	}
	/** When enabled, automatically create new mos-devices on-the-fly when receiving messages to unregistered MOS-ID:s */
	openRelay?:
		| boolean
		| {
				// options for on-the-fly-created connections
				options: IMOSDeviceConnectionOptions['primary']
		  }
}
export interface IMOSDeviceConnectionOptions {
	/** Connection options for the Primary NCS-server */
	primary: {
		/** Name (NCS ID) of the NCS-server */
		id: string
		/** Host address (IP-address) of the NCS-server  */
		host: string // ip-addr
		/** (Optional): Custom ports for communication */
		ports?: {
			upper: number
			lower: number
			query: number
		}
		/** (Optional) Timeout for commands (ms) */
		timeout?: number
		/** (Optional) Interval for sending of hearbeats (ms) */
		heartbeatInterval?: number
		/** (Optional) Some server doesn't expose the Query port, which can cause connection-errors.
		 * Set this to true to not use that port (will cause some methods to stop working)
		 */
		dontUseQueryPort?: boolean
	}
	/** Connection options for the Secondary (Buddy) NCS-server */
	secondary?: {
		/** Name (NCS ID) of the Buddy NCS-server */
		id: string
		/** Host address (IP-address) of the NCS-server  */
		host: string
		/** (Optional): Custom ports for communication */
		ports?: {
			upper: number
			lower: number
			query: number
		}
		/** (Optional) Timeout for commands (ms) */
		timeout?: number
		/** (Optional) Interval for sending of hearbeats (ms) */
		heartbeatInterval?: number

		/** (Optional) Some server doesn't expose the Query port, which can cause connection-errors.
		 * Set this to true to not use that port (will cause some methods to stop working)
		 */
		dontUseQueryPort?: boolean

		/** (Optional) Treat the secondary server as a "hot standby".
		 * A "hot standby" is a server that is powered down / in standby while the primary server is up.
		 * When a server is a hot standby it is expected to be disconnected and hence we will not send
		 * heartbeat messages. The connection status will still be reported as disconnected however as we
		 * do not pretend to be connected to something that is powered down.
		 *
		 * (This was added to support the hot standby feature of OpenMedia NRCS.)
		 */
		openMediaHotStandby?: boolean
	}
}